% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_overlap_pt_visualize.R
\name{circle_overlap_point_visualize}
\alias{circle_overlap_point_visualize}
\title{Visualize Circle Overlap Point}
\usage{
circle_overlap_point_visualize(
  circle1_center,
  circle1_radius,
  circle2_center,
  circle2_radius,
  min_spat_res
)
}
\arguments{
\item{circle1_center}{The center of the first circle as a dataframe with two columns one row}

\item{circle1_radius}{The radius of the first circle}

\item{circle2_center}{The center of the second circle as a dataframe with two columns one row}

\item{circle2_radius}{The radius of the second circle}

\item{min_spat_res}{The minimum spatial resolution of spatial variables. If sample space has a width less than this value, the middle of the overlap space is chosen as the point.}
}
\description{
This function returns a point uniformly sampled in the overlap of two circles and plots a visual aid of the circles the parallelogram sampled points that fail and the final point selected
}
\examples{
circle_overlap_point_visualize()
}
\keyword{modeling}
\keyword{movement}
\keyword{random}
\keyword{route}
\keyword{spatial}
