% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_routes.R
\name{coord_routes}
\alias{coord_routes}
\title{Generate Lists of Longitudes and Latitudes for Points Along Routes}
\usage{
coord_routes(
  start_long_lat,
  end_long_lat,
  max_travel_spd,
  step_time_interval,
  total_travel_time,
  min_spatial_res,
  number_of_routes
)
}
\arguments{
\item{start_long_lat}{The starting point as a longitude and latitude pair}

\item{end_long_lat}{The end point as a longitude and latitude pair}

\item{max_travel_spd}{The maximum traveling speed (be sure to use consitent distance and time units)}

\item{step_time_interval}{The time interval for each step between start and end point}

\item{total_travel_time}{The total amount of time it takes to get from start to end point (use consistend time units)}

\item{min_spatial_res}{The minimum spatial resolution of spatial variables. Passed to circle_overlap_point function}

\item{number_of_routes}{How many random routes you wish to generate}
}
\description{
This function returns lists of latitudes and longitudes for the points along routes randomly sampled between start and end points.
}
\examples{
coord_routes()
}
\keyword{modeling}
\keyword{movement}
\keyword{random}
\keyword{route}
\keyword{spatial}
